Remove uneeded casts

+++ Handlers/Power/Mac/MacPower.cs
@@ -26,7 +26,7 @@ namespace FOG.Handlers.Power
     {
         public void Shutdown(string comment, Power.FormOption options = Power.FormOption.Abort, string message = null, int seconds = 30)
         {
-            var minutes = ((double)seconds) / 60.0;
+            var minutes = seconds / 60.0;
             var timeDelay = (int)Math.Round(minutes);
 
             Power.QueueShutdown(string.Format("-h +{0} \"{1}\"", timeDelay, comment), options, message);
@@ -34,7 +34,7 @@ namespace FOG.Handlers.Power
 
         public void Restart(string comment, Power.FormOption options = Power.FormOption.Abort, string message = null, int seconds = 30)
         {
-            var minutes = ((double)seconds) / 60.0;
+            var minutes = seconds / 60.0;
             var timeDelay = (int)Math.Round(minutes);
 
             Power.QueueShutdown(string.Format("-r +{0} \"{1}\"", timeDelay, comment), options, message);
